Client-side routing for recent EMS versions
Does not yet support EMS <= v2
