+++ package.json
@@ -61,17 +61,21 @@
     "@mapbox/mapbox-gl-rtl-text": "^0.2.3",
     "@turf/bbox": "6.0.1",
     "@turf/center": "6.0.1",
+    "lodash": "^4.17.20",
     "mapbox-gl": "1.5.0",
     "markdown-it": "10.0.0",
     "moment": "2.24.0",
     "react": "16.11.0",
     "react-dom": "16.11.0",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "semver": "^7.3.2",
     "topojson-client": "3.0.1",
     "url-parse": "1.4.7",
     "whatwg-fetch": "^3.0.0"
   },
   "scripts": {
-    "dev": "webpack-dev-server --config webpack.dev.js",
+    "dev": "webpack-dev-server --config webpack.common.js",
     "lint": "eslint ./public/js --fix",
     "compile": "webpack --config webpack.prod.js",
     "build": "./bin/git-check.sh && yarn build-unsafe",
