+++ public/index.html
@@ -2,8 +2,9 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
     <title>Elastic Maps Service</title>
-    <link rel="shortcut icon" href="./ems.png" />
+    <link rel="shortcut icon" href="/ems.png" />
 </head>
 <body>
 <div id="wrapper"></div>
